Project ID - Int to Long #30

+++ acropolis-service/src/main/kotlin/org/ephyra/acropolis/service/api/ISystemSoftwareService.kt
@@ -3,7 +3,7 @@ package org.ephyra.acropolis.service.api
 import org.ephyra.acropolis.persistence.api.entity.SystemSoftwareEntity
 
 interface ISystemSoftwareService {
-    fun create(projectId: Int, name: String)
+    fun create(projectId : Long, name: String)
 
     fun get(name: String): SystemSoftwareEntity?
 }
